Fix: wrong url in nav bar
When both relative_url and root are configured in _config.yml, and menu in _config.mashiro.yml are configured as absolute paths
as below, URLs in nav bar have absolute paths without root being prepended.

@_config.yml
relative_link: true
root: /blog/

@_config.mashiro.yml
menu:
  Home: /
  Archives: /archives/
  About: /about/

In the example above, link is rendered as <a class="main-nav-link" href="/archives/">Archives</a>.
The reason is, relative_link config causes the path
configured in menu to be passed to relativeUrlHelper(),
which handles absolute paths wrongly.

We can blame this to our user to enable relative_link
while using absolute path in menu, but we can be more kind
to our users, by controlling who the path is passed to in
the nav bar ejs file. Specifically, if the path starts
with '/', treat as relative, otherwise treat as absolute,
no matter how relative_link is configured.